--- crates/tests/enumerators/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "test_enumerators"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[dependencies.windows-core]
-version = "0.59"
-
-[dependencies.microsoft-dia]
-path = "../../../"
