--- crates/tests/enumerators/src/lib.rs
@@ -1,1 +1,0 @@
-
