+++ tests/main.rs
@@ -1,0 +1,4 @@
+mod common;
+
+mod basic;
+mod enumerators;
